Use make+pandoc to generate the pages from markdown
Also, automate this with Travis.

This gets the output HTML pretty darn close to the current HTML; it jumps
through some silly hoops to make it very similar (using sed to use <tt>
tags instead of pandoc's default <code> tags, et c.).  If you diff them
ignoring space change, it's just a handful of lines that are different,
and you can verify that they are equivalent.